feat: get version number straight from codemeta.json
To avoid having to put version number and URL info into additional
files, I came up with a way to read them directly from the
codemeta.json file. It requires using GAS's facilities for reading
HTML files, which means we need the contents of codemeta.json in a
file that ends with a suffix of '.html'. The solution adopted here is
to have a symlink named 'version.html'. So, the code in Code.js now

1. reads version.html (which is really codemeta.json) using GAS's `HtmlService`
2. parses it as JSON to create an object
3. looks up the version number and other things as `.version` on the object