chore: set up 2 different clasp profiles
The reason for two separate project profiles is that we have to keep two
separate Apps Script projects on Google. The Caltech one is the real
development copy, but it's private (because it's inside the Library's
G-Suite/Google Workspace), so I keep a second copy in my own (mhucka)
Google account so that I can point people to it.