fix: htmlhint wasn't doing anything
I guess if it finds a `.htmlhintrc` file, it _only_ applies the rules
it finds in there; it doesn't apply all rules and override it with the
rc file values. So if you don't put anything in the file, nothing
happens. Not what I expected and not well explained anywhere, IMHO.