fix: detect when credentials need to be refreshed
At some point I removed a check for a valid token, and now what was
happening is that an operation would fail and the user would get an
error, instead of the original intention which was to bring up the
credentials dialog automatically. This does it now. It means it's
slower in some cases because it's doing an extra URL fetch, but I
think it's better to accept that in order to make the credentials
thing more user-friendly.