fix: get correct number of records expected
Previous code was an optimization to save one HTTP call, but led to
falling into the trap that when you get records and the number of
records exceeds 1000, Folio sometimes will report totalRecords = 1000
but that's an estimate, and there may be a lot more. The proper way is
to issue the query with limit = 0 first, and get the number of records
that way, before starting to get the records.